Change internal handling of include-*-list (#15)
We were re-splitting the list strings (include-file-list and include-function-list) for every given function in order to check whether it matches something in the list. Unfortunately, `strtok_r` "modifies" the string it is given, so it was pretty much invalidated after the first check. I think this explains #12.

Now we're splitting the list strings only once, at the beginning, and reusing it afterwards. I had to do some memory allocations, but it's only done once at startup and only depends on the CLI args, so it should be very reasonable.

This also reverts #14

Signed-off-by: Christophe Bedard <bedard.christophe@gmail.com>